Empecé a hacer el validate del GameForm

+++ PlataformaVideojuegos/src/main/java/Model/DTO/Game/GameDTO.java
@@ -10,7 +10,7 @@ public class GameDTO {
     private String developer;
     private LocalDate launchDate;
     private float basePrice;
-    private float currentDescount;
+    private int currentDescount;
     private GameCategory category;
     private AgeClasification ageClasification;
     private String[] availabeLanguages;
@@ -44,7 +44,7 @@ public class GameDTO {
         return basePrice;
     }
 
-    public float getCurrentDescount() {
+    public int getCurrentDescount() {
         return currentDescount;
     }
 
@@ -88,7 +88,7 @@ public class GameDTO {
     //Constructor
 
 
-    public GameDTO(Long id, String tittle, String description, String developer, LocalDate launchDate, float basePrice, float currentDescount, GameCategory category, AgeClasification ageClasification, String[] availabeLanguages, GameState state) {
+    public GameDTO(Long id, String tittle, String description, String developer, LocalDate launchDate, float basePrice, int currentDescount, GameCategory category, AgeClasification ageClasification, String[] availabeLanguages, GameState state) {
         this.id = id;
         this.tittle = tittle;
         this.description = description;
